requirement.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0
django-allauth==65.13.1
Faker==38.2.0
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.1
